+++ src/includes/traits/Facades/Image.php
@@ -11,10 +11,10 @@ namespace WebSharks\Core\Traits\Facades;
 use WebSharks\Core\Classes;
 use WebSharks\Core\Interfaces;
 use WebSharks\Core\Traits;
-#
+//
 use WebSharks\Core\Classes\Core\Error;
 use WebSharks\Core\Classes\Core\Base\Exception;
-#
+//
 use function assert as debug;
 use function get_defined_vars as vars;
 
@@ -38,6 +38,18 @@ trait Image
     }
 
     /**
+     * @since 17xxxx SVG converter.
+     *
+     * @param mixed ...$args Variadic args to underlying utility.
+     *
+     * @see Classes\Core\Utils\Image::svgToPng()
+     */
+    public static function svgToPng(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->©Image->svgToPng(...$args);
+    }
+
+    /**
      * @since 161010 PNG compression.
      *
      * @param mixed ...$args Variadic args to underlying utility.
@@ -47,5 +59,29 @@ trait Image
     public static function compressPng(...$args)
     {
         return $GLOBALS[static::class]->Utils->©Image->compressPng(...$args);
+    }
+
+    /**
+     * @since 17xxxx Geo-patterns.
+     *
+     * @param mixed ...$args Variadic args to underlying utility.
+     *
+     * @see Classes\Core\Utils\Image::geoPattern()
+     */
+    public static function geoPattern(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->©Image->geoPattern(...$args);
+    }
+
+    /**
+     * @since 17xxxx Geo-patterns.
+     *
+     * @param mixed ...$args Variadic args to underlying utility.
+     *
+     * @see Classes\Core\Utils\Image::geoPatternThumbnail()
+     */
+    public static function geoPatternThumbnail(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->©Image->geoPatternThumbnail(...$args);
     }
 }